remove `debug.*` prefix workaround to Title Case renaming quirk on Windows and Unix (part 2/2) for previously lowercase component filenames: - Box - Link - Typography
