+++ apps/emergence/app/page.tsx
@@ -1,6 +1,6 @@
 import { type ReactElement } from 'react'
 import { Page } from '@repo/mui/components/Page/Page'
-import { Secondary as Link } from '@repo/mui/components/link/Secondary'
+import { Secondary as Link } from '@repo/mui/components/Link/Secondary'
 
 export default function Home(): ReactElement {
   return (
