+++ packages/mui/src/components/Box/AlignedCenter.tsx
