+++ packages/mui/src/components/Box/AlignedRight.tsx
