+++ packages/mui/src/components/Box/Compact.tsx
