+++ packages/mui/src/components/ThemeSwitch.tsx
@@ -5,7 +5,7 @@ import InputLabel from '@mui/material/InputLabel'
 import MenuItem from '@mui/material/MenuItem'
 import Select, { type SelectChangeEvent } from '@mui/material/Select'
 import { useColorScheme } from '@mui/material/styles'
-import { AlignedRight } from './box/AlignedRight'
+import { AlignedRight } from './Box/AlignedRight'
 
 export function ThemeSwitch(): ReactElement | undefined {
   const { mode, setMode } = useColorScheme()
